+++ clang/lib/Frontend/CompilerInvocation.cpp
@@ -2589,6 +2589,10 @@ void CompilerInvocationBase::GenerateDiagnosticArgs(
     // This option is automatically generated from CheckConstexprFunctionBodies.
     if (Warning == "invalid-constexpr" || Warning == "no-invalid-constexpr")
       continue;
+    // This option is automatically generated from UnknownAttributeNamespaces.
+    if (Warning == "unknown-attribute-namespaces" &&
+        Opts.UnknownAttributeNamespaces.size())
+      continue;
     Consumer(StringRef("-W") + Warning);
   }